feat: icons for different history actions added
