+++ src/components/ActionIcon/index.ts
@@ -1,0 +1,2 @@
+import { ActionIcon } from './ActionIcon';
+export { ActionIcon };
