+++ src/redux/history/historySlice.ts
@@ -1,36 +1,8 @@
+import { LayerHistoryActions } from '@/types/historyTypes';
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
-
-export enum LayerHistoryActions {
-  Add,
-  Remove,
-  Rename,
-  ChangeOpacity,
-  ChangeVisibility,
-  ChangeOrder,
-}
-
-export function getHistoryName(action: LayerHistoryActions): string {
-  switch (action) {
-    case LayerHistoryActions.Add:
-      return 'Новый слой';
-    case LayerHistoryActions.Remove:
-      return 'Удаление слоя';
-    case LayerHistoryActions.Rename:
-      return 'Переименование слоя';
-    case LayerHistoryActions.ChangeOpacity:
-      return 'Непрозрачность слоя';
-    case LayerHistoryActions.ChangeVisibility:
-      return 'Видимость слоя';
-    case LayerHistoryActions.ChangeOrder:
-      return 'Порядок слоев';
-    default:
-      return '';
-  }
-}
 
 export type HistoryT = {
   id: number;
-  name: string;
   type: LayerHistoryActions;
 };
 
@@ -58,7 +30,6 @@ export const historySlice = createSlice({
     addNewHistoryItem: (state, action: PayloadAction<LayerHistoryActions>) => {
       const newAction: HistoryT = {
         id: state.historyIdCount,
-        name: 'Новое действие',
         type: action.payload,
       };
       state.historyIdCount++;
