+++ src/types/historyTypes.ts
@@ -1,0 +1,8 @@
+export enum LayerHistoryActions {
+  Add,
+  Remove,
+  Rename,
+  ChangeOpacity,
+  ChangeVisibility,
+  ChangeOrder,
+}
